refactor: rename components to fit naming scheme
